Updates to how "what song is this"-type intents work when allow_master_control is true
Some previous changes have been reverted to reduce changes when allow_master_control is false.

One change that applies regardless of the value of allow_master_control, though, is the addition of a call to enable_playing_intents() in the resume() function, as it seemed like the current functionality would mean that the "what song is this"-type intents would be disabled when the music was paused, and not re-enabled once the music is resumed (which is presumably the desired behaviour).